django/backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-oauth-toolkit==3.0.1
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
drf-social-oauth2==3.1.0
gunicorn==23.0.0
idna==3.10
jwcrypto==1.5.6
kombu==5.4.2
Markdown==3.7
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
